# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Mobil - OpenAPI 3.0",
    "description": "Dokumentasi untuk API Rental Mobil",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    },
    {
      "url": "http://rental-mobil.xyz"
    }
  ],
  "tags": [
    {
      "name": "mobil",
      "description": "Everything about your mobils"
    },
    {
      "name": "auth",
      "description": "Authentication"
    }
  ],
  "paths": {
    "/mobils": {
      "get": {
        "tags": [
          "mobil"
        ],
        "summary": "Get all mobil data",
        "description": "Listing all mobil data",
        "operationId": "getMobils",
        "responses": {
          "200": {
            "$ref": "#/components/responses/mobil"
          },
          "404": {
            "description": "Mobil not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/AddMobil": {
      "post": {
        "tags": [
          "mobil"
        ],
        "summary": "Add a new mobil to database",
        "description": "Add a new mobil to database",
        "operationId": "AddMobil",
        "requestBody": {
          "description": "Create a mobil in the database",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/mobil"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/mobil"
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/mobil/{id}": {
      "get": {
        "tags": [
          "mobil"
        ],
        "summary": "Find mobil by ID",
        "description": "Returns a single mobil",
        "operationId": "getMobil",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of motor to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/mobil"
          },
          "400": {
            "description": "Invalid ID "
          },
          "404": {
            "description": "Mobil not found"
          }
        }
      },
      "put": {
        "tags": [
          "mobil"
        ],
        "summary": "Updates a mobil in the store with form data",
        "description": "",
        "operationId": "updateMobil",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of mobil that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/mobil"
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "mobil"
        ],
        "summary": "Deletes a mobil",
        "description": "delete a mobil",
        "operationId": "deleteMobil",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Motor id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid id value"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Login",
        "operationId": "login",
        "requestBody": {
          "description": "Login",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password1!"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "berhasil login"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Register",
        "operationId": "register",
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Sodikul Wabil"
                  },
                  "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password1!"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "register berhasil"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "mobil": {
        "required": [
          "name",
          "description",
          "tipe",
          "harga_sewa",
          "foto",
          "manufactur"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Avanza"
          },
          "description": {
            "type": "string",
            "example": "Nyaman dan luas"
          },
          "tipe": {
            "type": "string",
            "example": "Medium"
          },
          "harga_sewa": {
            "type": "number",
            "example": 550000
          },
          "foto": {
            "type": "string",
            "example": "http://example.com/zx25r.jpg"
          },
          "manufactur": {
            "type": "string",
            "example": "2010"
          }
        }
      }
    },
    "requestBodies": {
      "mobil": {
        "description": "Mobil object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/mobil"
            }
          }
        }
      }
    },
    "responses": {
      "mobil": {
        "description": "Mobil object",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "OK"
                },
                "data": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/mobil"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 10
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updateAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}